+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [1.3.0] - 2020-10-15
+
+### Added
+
+- Occurrence enrollments with no guardians now have placeholder values
+- Number of people in queue and enrolled to event occurrence
+- Number of people in queue to event page
+
+### Fixed
+
+- Reloading occurrence show field no longer returns a blank page
+
 # 1.2.0
 
 - Add occurrence `capacityOverride` field handling in create, edit and show occurrence views